main/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.2.1
crispy-bootstrap5==2024.2
cryptography==42.0.7
cycler==0.12.1
Django==5.0.4
django-allauth==0.62.1
django-crispy-forms==2.1
django-taggit==5.0.1
django-tinymce==4.0.0
django-widget-tweaks==1.5.0
fabio==2024.4.0
fonttools==4.51.0
h5py==3.11.0
hdf5plugin==4.4.0
idna==3.7
Jinja2==3.1.4
kiwisolver==1.4.5
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib==3.9.0
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pillow==10.3.0
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.9.0.post0
requests==2.31.0
requests-oauthlib==2.0.0
silx==2.1.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
Whoosh==2.7.4
